feat: move sdk security
