+++ web/sdk/react/views/security/index.ts
@@ -1,0 +1,2 @@
+export { SecurityPage } from './security-page';
+export type { SecurityPageProps } from './security-page';
